+++ px_database_utility/cli.py
@@ -1,9 +1,9 @@
 '''Command line interface'''
 
-import sys
-import os
 import argparse
 import logging
+import os
+import sys
 
 log = logging.getLogger(__name__)
 
@@ -14,34 +14,39 @@ def get_cl_arguments():
     parser.add_argument("-o", "--operation", type=str, required=True,
                         choices=['BACKUP', 'RESTORE', 'LIST'],
                         help="Primary operations."
-                       )
+                        )
     parser.add_argument("-d", "--database", type=str, default='SELECT',
                         help="Specify single DB or all with 'ALL' | Restore: Specify the DB name to restore to."
-                       )
+                        )
     parser.add_argument("-i", "--interface", type=str,
                         help="Specify database host"
-                       )
+                        )
     parser.add_argument("-p", "--port", type=str,
                         help="Specify database post"
-                       )
+                        )
     parser.add_argument("-u", "--username", type=str,
                         help="Specify database username"
-                       )
+                        )
     parser.add_argument("-pw", "--password", type=str,
                         help="Specify database password"
-                       )
+                        )
     parser.add_argument("-f", "--file", type=str,
                         help="RESTORE only: Restore from absolute file path."
-                       )
+                        )
+    parser.add_argument("-k", "--keep", type=int, default=5,
+                        help="Number of backups to keep"
+                        )
     args = parser.parse_args()
 
     if args.operation == 'RESTORE':
         if args.database is None or args.database == 'SELECT':
-            log.error('You need to define a target name `--database DATABASE` for restore operations.')
+            log.error(
+                'You need to define a target name `--database DATABASE` for restore operations.')
             sys.exit(1)
         else:
             if args.database.endswith('.dmp'):
-                log.error('Use the --file flag for the absolute path; use the --database flag for the target database name.')
+                log.error(
+                    'Use the --file flag for the absolute path; use the --database flag for the target database name.')
                 sys.exit(1)
 
         if args.file is not None:
@@ -53,7 +58,6 @@ def get_cl_arguments():
                     log.error('File {} could not be found.'.format(args.file))
                     sys.exit(1)
 
-
     return {
         'operation': args.operation,
         'database': args.database,
@@ -61,5 +65,6 @@ def get_cl_arguments():
         'port': args.port,
         'username': args.username,
         'password': args.password,
-        'file': args.file
+        'file': args.file,
+        'keep': args.keep
     }